src/lambda1/pyproject.toml
[project]
name = "lambda1"
version = "0.1.0"
description = "Lambda function 1"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "openai>=1.0.0",
    "google-api-python-client>=2.110.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"